# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mahakali Engineers Client Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portal/projects/{id}/photos:
    get:
      operationId: listPhotos
      tags: [portal]
      summary: List project photos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
  /portal/projects/{id}/documents:
    get:
      operationId: listDocuments
      tags: [portal]
      summary: List project documents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
  /portal/projects/{id}/updates:
    get:
      operationId: listUpdates
      tags: [portal]
      summary: List project updates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Update"
  /portal/projects/{id}/payments:
    get:
      operationId: listPayments
      tags: [portal]
      summary: List project payments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
  /portal/projects/{id}/milestones:
    get:
      operationId: listMilestones
      tags: [portal]
      summary: List project milestones
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
  /portal/projects/{id}/comments:
    get:
      operationId: listComments
      tags: [portal]
      summary: List project comments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: addComment
      tags: [portal]
      summary: Add a comment to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
components:
  schemas:
...
    Photo:
      type: object
      required: [id, projectId, url, status, uploadedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        uploadedById:
          type: ["integer", "null"]
        url:
          type: string
        caption:
          type: ["string", "null"]
        status:
          type: string
        uploadedAt:
          type: string
...
    Document:
      type: object
      required: [id, projectId, name, url, type, status, uploadedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        uploadedById:
          type: ["integer", "null"]
        name:
          type: string
        url:
          type: string
        type:
          type: string
        status:
          type: string
        uploadedAt:
          type: string
...
    Comment:
      type: object
      required: [id, projectId, userId, message, authorName, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        userId:
          type: integer
        authorName:
          type: string
        message:
          type: string
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
    Update:
      type: object
      required: [id, projectId, message, postedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        message:
          type: string
        postedAt:
          type: string
...
    Milestone:
      type: object
      required: [id, projectId, title, order, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Payment:
      type: object
      required: [id, projectId, label, amount, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        label:
          type: string
        amount:
          type: string
        status:
          type: string
        dueDate:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
...