# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mahakali Engineers Client Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  projectId: number;
  userId: number;
  authorName: string;
  message: string;
  createdAt: string;
}